Make the default to not create VPC
As we are moving towards one shared VPC for all our clusters (per
environment), just for code tidyness, we want to make it the default
behaviour of the module to NOT create the VPC, and instead reference the
VPC via a data source. This still eanbles creating a VPC when needed,
but going forward this module will be used more for creating subnets in
the shared VPCs.